Applied PR comments.
